ship chair healing reimplementation + bed fixes
dunno if you guys wanna keep this, but this was mentioned in the original mod description on the SB site, figured i'd try to reimplment it.

the ship chair in peglaci ships now provides regeneration when sat in. it also stops hunger, similar to the bed resting effect. UNLIKE the resting effect, it won't strip you naked when you sit in it :)

seems to work fine, but again, up to you guys if you wish to keep it or tweak the regen rate. at the moment, it only uses t1 healing.

additionally, peglaci beds now provide the resting buff instead of both regen and nude. resting p much has both these effects, but might as well have it for the sake of consistency with vanilla beds.